Don't require `@unknown default` for resilient public or package enums if accessed by other modules in the same package.
It's generally required for resilient enums that might be modified in
the future and the accessing module is not re-built with that change;
however, this is not the case for modules within the same package as
they are required to be built together. Thus we can skip the diagnostics
for adding `@unknown default` at the switch use site.

The PR only modifies the diagnostics, i.e. does not affect SILGen.

Resolves rdar://130015149.
